monitoring/instrument/metrics/requirements.txt
opentelemetry-api==1.19.0
opentelemetry-sdk==1.19.0
opentelemetry-instrumentation-asgi==0.40b0
opentelemetry-instrumentation-fastapi==0.40b0
opentelemetry-instrumentation-requests==0.40b0
opentelemetry-instrumentation-logging==0.40b0
opentelemetry-exporter-jaeger==1.19.0
opentelemetry-exporter-otlp-proto-grpc==1.19.0
opentelemetry-exporter-prometheus==1.12.0rc1
fastapi==0.96.0
python-multipart==0.0.6
prometheus-client==0.17.1
loguru
numpy==1.23.1
opencv-python
transformers==4.44.2
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
imagehash
Pillow
uvicorn
pytest